Updates to Import
